chore: Add publish tool
